Fix process unit tests
